pyproject.toml
[tool.poetry]
name = "bravetogether-backend"
version = "0.1.0"
description = ""
authors = ["Philipp Page <github@philipp-page.de>"]
readme = "README.md"
packages = [{include = "bravetogether_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.2"
flask-cors = "^3.0.10"
flask-pymongo = "^2.3.0"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^6.0.0"
mypy = "^0.991"
types-flask-cors = "^3.0.10"
pymongo-stubs = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
files = "python_starter"
